# perf(db): use pk id for lookups and ordering