--- a/internal/database/postgres.go
+++ b/internal/database/postgres.go
@@ -231,14 +231,20 @@ func (db *PostgreSQL) UpdateServer(ctx context.Context, id string, server *apiv0
 	}
 
 	// Marshal updated server
-	// keep the id invariant: the json meta id must match the primary key id
+	// fail fast if a request attempts to change the id in the json payload
+	if server.Meta != nil && server.Meta.Official != nil && server.Meta.Official.ID != "" && server.Meta.Official.ID != id {
+		return nil, fmt.Errorf("%w: meta.official.id (%s) must match path id (%s)", ErrInvalidInput, server.Meta.Official.ID, id)
+	}
+	// ensure json meta exists and set id if missing (but do not overwrite if present)
 	if server.Meta == nil {
 		server.Meta = &apiv0.ServerMeta{}
 	}
 	if server.Meta.Official == nil {
 		server.Meta.Official = &apiv0.RegistryExtensions{}
 	}
-	server.Meta.Official.ID = id
+	if server.Meta.Official.ID == "" {
+		server.Meta.Official.ID = id
+	}
 
 	valueJSON, err := json.Marshal(server)
 	if err != nil {