--- a/internal/database/postgres.go
+++ b/internal/database/postgres.go
@@ -94,15 +94,15 @@ func (db *PostgreSQL) List(
 		}
 	}
 
-	// Add cursor pagination using registry metadata ID
-	if cursor != "" {
-		if _, err := uuid.Parse(cursor); err != nil {
-			return nil, "", fmt.Errorf("invalid cursor format: %w", err)
-		}
-		whereConditions = append(whereConditions, fmt.Sprintf("(value->'_meta'->'io.modelcontextprotocol.registry/official'->>'id') > $%d", argIndex))
-		args = append(args, cursor)
-		argIndex++
-	}
+    // Add cursor pagination using primary key ID
+    if cursor != "" {
+        if _, err := uuid.Parse(cursor); err != nil {
+            return nil, "", fmt.Errorf("invalid cursor format: %w", err)
+        }
+        whereConditions = append(whereConditions, fmt.Sprintf("id > $%d", argIndex))
+        args = append(args, cursor)
+        argIndex++
+    }
 
 	// Build the WHERE clause
 	whereClause := ""
@@ -111,13 +111,13 @@ func (db *PostgreSQL) List(
 	}
 
 	// Simple query on servers table
-	query := fmt.Sprintf(`
-		SELECT value
-		FROM servers
-		%s
-		ORDER BY (value->'_meta'->'io.modelcontextprotocol.registry/official'->>'id')
-		LIMIT $%d
-	`, whereClause, argIndex)
+    query := fmt.Sprintf(`
+        SELECT value
+        FROM servers
+        %s
+        ORDER BY id
+        LIMIT $%d
+    `, whereClause, argIndex)
 	args = append(args, limit)
 
 	rows, err := db.conn.Query(ctx, query, args...)
@@ -168,7 +168,7 @@ func (db *PostgreSQL) GetByID(ctx context.Context, id string) (*apiv0.ServerJSON
 	query := `
 		SELECT value
 		FROM servers
-		WHERE (value->'_meta'->'io.modelcontextprotocol.registry/official'->>'id') = $1
+		WHERE id = $1
 	`
 
 	var valueJSON []byte
@@ -240,7 +240,7 @@ func (db *PostgreSQL) UpdateServer(ctx context.Context, id string, server *apiv0
 	query := `
 		UPDATE servers 
 		SET value = $1
-		WHERE (value->'_meta'->'io.modelcontextprotocol.registry/official'->>'id') = $2
+		WHERE id = $2
 	`
 
 	result, err := db.conn.Exec(ctx, query, valueJSON, id)